#181-Reorganize Tests directory structure
Done!
