--- src/Tests/Scanner_Tokens/Data/tokenization_compound_tokens.ty
@@ -1,24 +1,0 @@
-// Compound tokens -- tokenization test source
-/***
-multi-lines comment
-***/
-@ @=
-& &=
-| |=
-^ ^^ ^= ^^=
-: :: ::=
-/ /=
-= ==
-! !! != !!=
-? ?? ?= ??=
-> >> >>> >= >>= >>>= >< ><=
-< << <<< <= <<= <<<=
-- -- -= ->
-% %=
-+ ++ +=
-* ** *= **=
-:= --= ++= <-
-. ... ..a
-    ///* */
- {} {{ }}
-
