--- src/Tests/Scanner_Tokens/Data/tokenization_simple_tokens.ty
@@ -1,8 +1,0 @@
-// Simple tokens -- tokenization test source
-[]
-()
-,;
-#
-.
-~
-?
